Move copy_basic_node_props() to node_utils.py
